requirements.txt
# Clawnbot Dependencies
anthropic>=0.18.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

